+++ src/wads2.cc
@@ -469,7 +469,7 @@ wf->directory = (DirPtr) GetMemory ((long) sizeof (struct Directory)
 if (fseek (wf->fp, wf->dirstart, SEEK_SET) != 0)
    {
    printf ("%.128s: can't seek to directory at %08lXh\n",
-      filename, wf->dirstart);
+	   filename, (unsigned long) wf->dirstart);
    fail = true;
    goto byebye;
    }
@@ -519,7 +519,7 @@ for (dir = MasterDir; dir; dir = dir->next)
    strncpy (dataname, dir->dir.name, WAD_NAME);
    fprintf (file, "%-*s  %-50s  %6ld  x%08lx\n",
     WAD_NAME, dataname, dir->wadfile->pathname (),
-    dir->dir.size, dir->dir.start);
+    (long) dir->dir.size, (unsigned long) dir->dir.start);
    if (file == stdout && lines++ > screen_lines - 4)
       {
       lines = 0;
@@ -556,9 +556,9 @@ for (n = 0; n < wad->dirsize; n++)
    strncpy (dataname, wad->directory[n].name, WAD_NAME);
    fprintf (file, "%-*s  %6ld  x%08lx  x%08lx\n",
      WAD_NAME, dataname,
-     wad->directory[n].size,
-     wad->directory[n].start,
-     wad->directory[n].size + wad->directory[n].start - 1);
+     (long) wad->directory[n].size,
+     (unsigned long) wad->directory[n].start,
+     (unsigned long) (wad->directory[n].size + wad->directory[n].start - 1));
    if (file == stdout && lines++ > screen_lines - 4)
       {
       lines = 0;
@@ -683,7 +683,7 @@ for (MDirPtr entry = MasterDir; entry != 0; entry = entry->next)
    strncpy (dataname, entry->dir.name, WAD_NAME);
    dataname[WAD_NAME] = '\0';
    fprintf (file, "Contents of entry %s (size = %ld bytes):\n",
-      dataname, entry->dir.size);
+	    dataname, (long) entry->dir.size);
    const Wad_file *wf = entry->wadfile;
    wf->seek (entry->dir.start);
    for (n = 0; n < entry->dir.size;)
